camera: allow loading files with text relocations from /system
the qcom camera driver needs to load a shared library from /system
which contains a text relocation. Allow it.

Due to bug 20013628, SELinux policy was inappropriately treating
an execmod denial as an execmem denial. Move to using a proper
execmod denial and get rid of execmem.

Addresses the following denial:
  [    3.701696] type=1400 audit(32272583.250:5): avc: denied { execmod } for pid=205 comm="mm-qcamera-daem" path="/system/vendor/lib/libmmcamera_faceproc.so" dev="mmcblk0p25" ino=1727 scontext=u:r:camera:s0 tcontext=u:object_r:system_file:s0 tclass=file

Bug: 20013628
Change-Id: I6a44c097383458fdbba49bc8756660052b2c395e